pyproject.toml
[project]
name = "inpdfile"
version = "0.1.0"
description = ""
authors = [
    {name = "Keniden",email = "nik.glumov.94@mail.ru"}
]
readme = "README.md"
requires-python = "3.12.7"
dependencies = [
    "alembic (==1.13.3)",
    "amqp (==5.3.1)",
    "annotated-types (==0.7.0)",
    "anyio (==4.8.0)",
    "apscheduler (==3.11.0)",
    "asyncpg (==0.30.0)",
    "bcrypt (==4.0.1)",
    "billiard (==4.2.1)",
    "celery (==5.4.0)",
    "certifi (==2024.12.14)",
    "click (==8.1.8)",
    "click-didyoumean (==0.3.1)",
    "click-plugins (==1.1.1)",
    "click-repl (==0.3.0)",
    "colorama (==0.4.6)",
    "dnspython (==2.7.0)",
    "ecdsa (==0.19.0)",
    "email-validator (==2.2.0)",
    "fastapi (==0.115.0)",
    "fastapi-cli (==0.0.7)",
    "greenlet (==3.1.1)",
    "gunicorn (==23.0.0)",
    "h11 (==0.14.0)",
    "httpcore (==1.0.7)",
    "httptools (==0.6.4)",
    "httpx (==0.28.1)",
    "idna (==3.10)",
    "iniconfig (==2.0.0)",
    "itsdangerous (==2.2.0)",
    "jinja2 (==3.1.4)",
    "kombu (==5.4.2)",
    "loguru (==0.7.2)",
    "mako (==1.3.8)",
    "markdown-it-py (==3.0.0)",
    "markupsafe (==3.0.2)",
    "mdurl (==0.1.2)",
    "motor (==3.7.0)",
    "orjson (==3.10.13)",
    "packaging (==24.2)",
    "passlib (==1.7.4)",
    "pillow (==11.1.0)",
    "pluggy (==1.5.0)",
    "prometheus-fastapi-instrumentator (==7.0.2)",
    "prometheus-client (==0.21.1)",
    "prompt-toolkit (==3.0.50)",
    "pyasn1 (==0.6.1)",
    "pydantic (==2.9.2)",
    "pydantic-extra-types (==2.10.1)",
    "pydantic-settings (==2.5.2)",
    "pydantic-core (==2.23.4)",
    "pygments (==2.19.1)",
    "pymongo (==4.11.1)",
    "pytest (==8.3.4)",
    "pytest-asyncio (==0.25.3)",
    "python-dateutil (==2.9.0.post0)",
    "python-dotenv (==1.0.1)",
    "python-jose (==3.3.0)",
    "python-json-logger (==3.2.1)",
    "python-multipart (==0.0.20)",
    "pyyaml (==6.0.2)",
    "redis (==5.2.1)",
    "rich (==13.9.4)",
    "rich-toolkit (==0.12.0)",
    "rsa (==4.9)",
    "shellingham (==1.5.4)",
    "six (==1.17.0)",
    "sniffio (==1.3.1)",
    "sqlalchemy (==2.0.35)",
    "starlette (==0.38.6)",
    "typer (==0.15.1)",
    "typing-extensions (==4.12.2)",
    "tzdata (==2025.1)",
    "tzlocal (==5.3)",
    "ujson (==5.10.0)",
    "uvicorn (==0.31.0)",
    "vine (==5.1.0)",
    "watchfiles (==1.0.3)",
    "wcwidth (==0.2.13)",
    "websockets (==14.1)",
    "win32-setctime (==1.2.0)",
    "autoflake (==2.3.1,<3.0.0)",
    "flake8 (==7.1.2,<8.0.0)",
    "pylint (==3.3.4,<4.0.0)",
    "mypy (==1.15.0,<2.0.0)",
    "flower (==2.0.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
